app/requirements.txt
flask
flask-swagger-ui
mysql-connector-python
flask-jwt-extended
requests